Changed code to use keryring module to avoid storing passwords in plain text
Keyring module integrates with best backend available on each operating
system like win32 crypto api on windows, gnome-keyring/kwallet on linux
to store credentials securely for applications.

So password is added
first time when it is read from configuration file and stored in keyring
and retrieved everytime from keyring.

If password is not present, it is prompted for the same and then stored
and retrieved.

Also if user wants to update password, he can provide -u option, which
will ignore the existing password and reads from file, if config file is
provided or prompts for the same, if not present in file and then sets
the password.