Update steps for sudo to prompt for password
Updated step 3 of both configurations to include -b (background flag) for running tincan. This way, the ipop-tincan process is put into background only after asking the password from the user. 
Note that doing this way, doesn't show up the process by running 'jobs' command on terminal. But killing it could be done in the same way as before.